feat(blockchain): Implement conversion from `CompiledModule` for Move module ABI response structure
